Make printing detect state changes.
By checking whether the state has changed, we can make sure the output
always shows a states condition. This way we won't miss a state even
though it takes fewer than 1000 cycles to complete. To make this
complete we need to move the printing below the state machine so that
it's output is up to date with the current state.